organize minari dataset
