+++ minari/__init__.py
@@ -1,6 +1,6 @@
 from minari.data_collector import DataCollectorV0
 from minari.data_collector.callbacks import EpisodeMetadataCallback, StepDataCallback
-from minari.minari_dataset import MinariDataset
+from minari.dataset.minari_dataset import MinariDataset
 from minari.storage.hosting import (
     download_dataset,
     list_remote_datasets,
@@ -13,5 +13,24 @@ from minari.utils import (
     create_dataset_from_collector_env,
 )
 
+
+__all__ = [
+    # Minari Dataset
+    "MinariDataset",
+    # Data collection
+    "DataCollectorV0",
+    "EpisodeMetadataCallback",
+    "StepDataCallback",
+    # Dataset Functions
+    "download_dataset",
+    "list_remote_datasets",
+    "upload_dataset",
+    "delete_dataset",
+    "list_local_datasets",
+    "load_dataset",
+    "combine_datasets",
+    "create_dataset_from_buffers",
+    "create_dataset_from_collector_env",
+]
 
 __version__ = "0.3.0"
