+++ minari/dataset/minari_dataset.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import os
 from dataclasses import dataclass
-from typing import Callable, Iterable, List, NamedTuple, Optional, Union
+from typing import Callable, Dict, Iterable, List, NamedTuple, Optional, Union
 
 import gymnasium as gym
 import h5py
@@ -10,8 +10,33 @@ import numpy as np
 from gymnasium.envs.registration import EnvSpec
 
 from minari.data_collector import DataCollectorV0
-from minari.minari_storage import MinariStorage, PathLike
-from minari.utils import clear_episode_buffer
+from minari.dataset.minari_storage import MinariStorage, PathLike
+
+
+def clear_episode_buffer(episode_buffer: Dict, eps_group: h5py.Group) -> h5py.Group:
+    """Save an episode dictionary buffer into an HDF5 episode group recursively.
+
+    Args:
+        episode_buffer (dict): episode buffer
+        eps_group (h5py.Group): HDF5 group to store the episode datasets
+
+    Returns:
+        episode group: filled HDF5 episode group
+    """
+    for key, data in episode_buffer.items():
+        if isinstance(data, dict):
+            if key in eps_group:
+                eps_group_to_clear = eps_group[key]
+            else:
+                eps_group_to_clear = eps_group.create_group(key)
+            clear_episode_buffer(data, eps_group_to_clear)
+        else:
+            # assert data is numpy array
+            assert np.all(np.logical_not(np.isnan(data)))
+            # add seed to attributes
+            eps_group.create_dataset(key, data=data, chunks=True)
+
+    return eps_group
 
 
 class EpisodeData(NamedTuple):
